Improve heuristic of phase 1.
For 200 cubes, we go from 23s initialization + 40s solving
to 34s initialization + 30s solving. (25% gain in pure solving time)

`flatIndex b` might benefit from a smaller base. That is an unmeasurable
improvement but the code ends up looking more uniform as in each phase
the same coordinate is now chosen for the base in every table.